+++ python/asteroid_prospector/__init__.py
@@ -1,11 +1,14 @@
 from .constants import N_ACTIONS, OBS_DIM
 from .hello_env import HelloProspectorEnv
+from .native_core import NativeCoreConfig, NativeProspectorCore
 from .reference_env import ProspectorReferenceEnv, ReferenceEnvConfig
 
 __all__ = [
     "HelloProspectorEnv",
     "ProspectorReferenceEnv",
     "ReferenceEnvConfig",
+    "NativeProspectorCore",
+    "NativeCoreConfig",
     "OBS_DIM",
     "N_ACTIONS",
 ]